Add optional workspace ID setting
By default, Toggl creates one Workspace for a user. Some users have
multiple workspaces, however, so the user should be able to specify the
workspace ID for the workspace they want to use toggl.vim with. Now they
can optionally specify that workspace ID with `let g:toggl_workspace_id = 987654321`
and toggl.vim will use that workspace. It does not verify that the
workspace exists or that the user has permission to write to that
workspace.